Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main()
 {
-    int a;
+    float a;
     scanf("%d", &a);
     print("Area: %.2f\n", 3.14 * a * a);
     return 0;
